requirements.txt
# SpoofSniff Project Dependencies
scapy
dnspython